api/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# AI/ML
google-generativeai>=0.3.0
youtube-transcript-api>=0.6.0

# PDF処理
PyMuPDF>=1.23.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
